Move meshModels in Front and Back folders
