+++ visualization/app/codeCharta/services/3DExports/3DPreview/MeshModels/FrontMeshModels/frontPrintContainerMesh.ts
@@ -1,6 +1,6 @@
-import { GeometryOptions } from "../preview3DPrintMesh"
-import { GeneralMesh, GeneralSizeChangeMesh } from "./generalMesh"
-import { BackPrintColorChangeStrategy } from "../ColorChangeStrategies/backPrintColorChangeStrategy"
+import { GeometryOptions } from "../../preview3DPrintMesh"
+import { GeneralMesh, GeneralSizeChangeMesh } from "../generalMesh"
+import { BackPrintColorChangeStrategy } from "../../ColorChangeStrategies/backPrintColorChangeStrategy"
 import { Font } from "three"
 import { FrontTextMesh } from "./frontTextMesh"
 import { SecondRowTextMesh } from "./secondRowTextMesh"
